notes/note_upload_service.py
from concurrent import futures
import grpc
import notes_pb2
import notes_pb2_grpc

class NoteUploadServiceServicer(notes_pb2_grpc.NoteUploadServiceServicer):
    def UploadNote(self, request, context):
        # Implement logic to handle the note upload, including saving the file to S3
        return notes_pb2.UploadNoteResponse(noteId='generatedNoteId')

def serve():
    server = grpc.server(futures.ThreadPoolExecutor(max_workers=10))
    notes_pb2_grpc.add_NoteUploadServiceServicer_to_server(NoteUploadServiceServicer(), server)
    server.add_insecure_port('[::]:50051')
    server.start()
    server.wait_for_termination()

if __name__ == '__main__':
    serve()
