notes/requirements.txt
grpcio
grpcio-tools
boto3